Added "verify_session_signature"
With the new Facebook Connect JavaScript SDK the format of the session
object has changed. The attributes are no longer prefaced with fb_ and
this changes the calculation of the signature.
See "Has the Cookie format changed" in the Facebook Connect JavaScript
SDK FAQ: http://wiki.github.com/facebook/connect-js/faq